+++ src/Filament/Forms/Components/ImageLibrarySelectExisting.php
@@ -4,6 +4,7 @@ namespace Outerweb\FilamentImageLibrary\Filament\Forms\Components;
 
 use Closure;
 use Filament\Forms\Components\Field;
+use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Pagination\LengthAwarePaginator;
 use Illuminate\Support\Arr;
 use Outerweb\ImageLibrary\Facades\ImageLibrary;
@@ -16,41 +17,45 @@ class ImageLibrarySelectExisting extends Field
 
     protected bool|Closure $multiple = false;
 
-    public function multiple(bool|Closure $multiple = true): static
+    public function multiple(bool|Closure $multiple = true) : static
     {
         $this->multiple = $multiple;
 
         return $this;
     }
 
-    public function itemsPerPage(int $perPage): static
+    public function itemsPerPage(int $perPage) : static
     {
         $this->itemsPerPage = $perPage;
 
         return $this;
     }
 
-    public function getAllowsMultiple(): bool
+    public function getItemsPerPage() : int
+    {
+        return $this->itemsPerPage;
+    }
+
+    public function getAllowsMultiple() : bool
     {
         return (bool) $this->evaluate($this->multiple);
     }
 
-    public function getImages(): LengthAwarePaginator
+    public function getImages() : Collection
     {
-        return ImageLibrary::imageModel()::latest()
-            ->paginate($this->itemsPerPage);
+        return ImageLibrary::imageModel()::latest()->get();
     }
 
-    public function getState(): mixed
+    public function getState() : mixed
     {
         return Arr::wrap(parent::getState() ?? []);
     }
 
-    protected function setUp(): void
+    protected function setUp() : void
     {
         parent::setUp();
 
-        $this->afterStateHydrated(static function (self $component, $state): void {
+        $this->afterStateHydrated(static function (self $component, $state) : void {
             $component->state($state ?? []);
         });
     }